+++ CHANGELOG.md
@@ -4,7 +4,15 @@
 
 ## [Unreleased]
 
-- support for IntelliJ IDEA 2024.3
+## [0.0.3] - 2026-04-17
+- Automatic detection of affected test files.
+- Added support for Kotlin tests.
+- Test execution cancellation.
+- Removed compatibility limit.
+
+## [0.0.2] - 2024-11-28
+
+- Support for IntelliJ IDEA 2024.3
 
 ## [0.0.1] - 2024-10-21